Move kinesis stream into Kinesis namespace
